Improve react components design
#26
