+++ frontend/src/components/bankAccountMainWidget/bankAccountMainWidget.js
@@ -4,42 +4,16 @@ import Grid from "@mui/material/Grid";
 import ButtonSelector from "../../utils/buttonSelector";
 import TransactionsTable from "../transactionsTable/transactionsTable";
 import BankContext from "../../contexts/BankContext";
-import {ClickAwayListener, Grow, MenuList, Paper, Popper, TextField, ThemeProvider} from "@mui/material";
-import {AdapterDayjs} from '@mui/x-date-pickers/AdapterDayjs';
-import {LocalizationProvider} from '@mui/x-date-pickers/LocalizationProvider';
-import {DatePicker} from '@mui/x-date-pickers/DatePicker';
-import NotificationsContext from "../../contexts/NotificationsContext";
+import PeriodSelector from "../../utils/periodSelector";
 
 function BankAccountMainWidget() {
-    const {theme, datePickerTheme} = useContext(CustomThemeContext);
-    const {enqueueErrorSnackbar} = useContext(NotificationsContext);
+    const {theme} = useContext(CustomThemeContext);
     const {
         bankAccount,
         getTransactions,
         transactions,
     } = useContext(BankContext);
-    const periodOptions = [
-        {
-            name: 'month',
-            disable: true,
-            onClick: () => setPeriod('month'),
-        },
-        {
-            name: 'year',
-            disable: true,
-            onClick: () => setPeriod('year'),
-        },
-        {
-            name: 'custom',
-            disable: false,
-            onClick: (event) => {
-                setOpenDateRangeSelector(true);
-                setAnchorEl(event.currentTarget);
-                handleDateRangeChange(dateRange);
-            },
-        },
-    ];
-    const [period, setPeriod] = useState(periodOptions[0].name);
+    const [period, setPeriod] = useState(null);
     const overviewOptions = [
         {
             name: 'transactions',
@@ -48,45 +22,15 @@ function BankAccountMainWidget() {
         }
     ];
     const [overview, setOverview] = useState(overviewOptions[0].name);
-    const [anchorEl, setAnchorEl] = useState(null);
-    const [openDateRangeSelector, setOpenDateRangeSelector] = useState(false);
-    const [dateRange, setDateRange] = useState({
-        startDate: null,
-        endDate: null,
-    });
-
-    const handleDateRangeChange = (newDateRange) => {
-        if (newDateRange.startDate && newDateRange.endDate) {
-            if (newDateRange.startDate > newDateRange.endDate) {
-                enqueueErrorSnackbar('Please make sure the start date is before the end date.');
-            } else {
-                setPeriod('custom');
-                getTransactions({
-                    bankAccountId: bankAccount.id,
-                    period: 'custom',
-                    dateRange: newDateRange,
-                });
-            }
-        }
-    }
 
     useEffect(() => {
-        if (bankAccount.id != null) {
+        if (bankAccount.id != null && period != null) {
             getTransactions({
                 bankAccountId: bankAccount.id,
                 period: period,
             });
         }
     }, [bankAccount.id]); // eslint-disable-line react-hooks/exhaustive-deps
-
-    useEffect(() => {
-        if (bankAccount.id != null) {
-            getTransactions({
-                bankAccountId: bankAccount.id,
-                period: period,
-            });
-        }
-    }, [period]); // eslint-disable-line react-hooks/exhaustive-deps
 
     return (
         <div
@@ -109,75 +53,19 @@ function BankAccountMainWidget() {
                 </Grid>
                 <Grid item xs={9}/>
                 <Grid item xs={2}>
-                    <ButtonSelector
-                        options={periodOptions}
-                        currentOption={period}
-                        setCurrentOption={setPeriod}
+                    <PeriodSelector
+                        period={period}
+                        setPeriod={setPeriod}
+                        action={(period, dateRange) => {
+                            if (bankAccount.id != null) {
+                                getTransactions({
+                                    bankAccountId: bankAccount.id,
+                                    period: period,
+                                    dateRange: dateRange,
+                                });
+                            }
+                        }}
                     />
-                    <Popper
-                        open={openDateRangeSelector}
-                        anchorEl={anchorEl}
-                        transition
-                        disablePortal
-                    >
-                        {({TransitionProps}) => (
-                            <Grow
-                                {...TransitionProps}
-                                style={{transformOrigin: 'right top'}}
-                            >
-                                <Paper>
-                                    <ClickAwayListener
-                                        onClickAway={() => {
-                                            setOpenDateRangeSelector(false);
-                                        }}
-                                    >
-                                        <MenuList
-                                            autoFocusItem={openDateRangeSelector}
-                                            style={{
-                                                padding: 10,
-                                            }}
-                                        >
-                                            <ThemeProvider theme={datePickerTheme}>
-                                                <LocalizationProvider dateAdapter={AdapterDayjs}>
-                                                    <DatePicker
-                                                        label='Start Date'
-                                                        value={dateRange.startDate}
-                                                        onChange={(newStartDate) => {
-                                                            setDateRange(prevDateRange => {
-                                                                const newDateRange = {
-                                                                    ...prevDateRange,
-                                                                    startDate: newStartDate,
-                                                                };
-                                                                handleDateRangeChange(newDateRange);
-                                                                return newDateRange;
-                                                            });
-                                                        }}
-                                                        renderInput={(params) => <TextField {...params}/>}
-                                                    />
-                                                    <DatePicker
-                                                        label='End Date'
-                                                        value={dateRange.endDate}
-                                                        onChange={(newEndDate) => {
-                                                            setDateRange(prevDateRange => {
-                                                                const newDateRange = {
-                                                                    ...prevDateRange,
-                                                                    endDate: newEndDate,
-                                                                };
-                                                                handleDateRangeChange(newDateRange);
-                                                                return newDateRange;
-                                                            });
-                                                        }}
-                                                        renderInput={(params) => <TextField {...params}
-                                                                                            sx={{svg: theme.palette.datePickerIcon}}/>}
-                                                    />
-                                                </LocalizationProvider>
-                                            </ThemeProvider>
-                                        </MenuList>
-                                    </ClickAwayListener>
-                                </Paper>
-                            </Grow>
-                        )}
-                    </Popper>
                 </Grid>
                 <Grid item xs={12}>
                     <TransactionsTable
