+++ frontend/src/components/dialogs/categoriesDialog/categoriesDialogTable.js
