+++ frontend/src/components/dialogs/dialogTemplate.js
