+++ frontend/src/components/dialogs/dialogTemplateActions.js
