+++ frontend/src/components/dialogs/dialogTemplateContent.js
