+++ frontend/src/components/dialogs/dialogTemplateTitle.js
