+++ frontend/src/components/dialogs/formatCategoryDialog/editCategoryDialogContent.js
@@ -1,12 +1,17 @@
 import Grid from "@mui/material/Grid";
 import {MenuItem, Select, TextField} from "@mui/material";
-import {useContext} from "react";
-import UtilsContext from "../../../../contexts/UtilsContext";
+import {createElement, useContext} from "react";
+import UtilsContext from "../../../contexts/UtilsContext";
 
-function AddBankAccountDialogContent({
-                                         onInputChange,
-                                     }) {
-    const {currencies} = useContext(UtilsContext);
+function EditCategoryDialogContent({
+                                       editCategory,
+                                       onInputChange,
+                                   }) {
+    const {MUIIcons} = useContext(UtilsContext);
+
+    const getIconName = (iconId) => {
+        return MUIIcons.filter(MUIIcon => MUIIcon.id === iconId)[0].name;
+    }
 
     return (
         <Grid container spacing={2} pl={2} pt={2} pr={2}>
@@ -21,13 +26,14 @@ function AddBankAccountDialogContent({
             >
                 <TextField
                     label="Name"
-                    name="name"
+                    name='name'
                     variant="standard"
                     style={{
                         marginLeft: 15,
                         width: '50%',
                     }}
                     onChange={onInputChange}
+                    defaultValue={editCategory ? editCategory.name : ''}
                 />
             </Grid>
             <Grid item xs={3}>
@@ -38,59 +44,43 @@ function AddBankAccountDialogContent({
             <Grid item xs={9}>
                 <TextField
                     label="Description"
-                    name="description"
+                    name='description'
                     multiline
                     rows={3}
                     maxRows={3}
                     fullWidth
                     onChange={onInputChange}
+                    defaultValue={editCategory ? editCategory.description : ''}
                 />
             </Grid>
             <Grid item xs={3}>
-                <div style={{marginTop: 20, fontWeight: 'bold'}}>
-                    IBAN:
+                <div style={{marginTop: 10, fontWeight: 'bold'}}>
+                    Icon:
                 </div>
-            </Grid>
-            <Grid
-                item
-                xs={9}
-            >
-                <TextField
-                    label="IBAN"
-                    name="IBAN"
-                    variant="standard"
-                    style={{
-                        marginLeft: 15,
-                        width: '50%',
-                    }}
-                    onChange={onInputChange}
-                />
             </Grid>
             <Grid item xs={3}>
-                <div style={{marginTop: 20, fontWeight: 'bold'}}>
-                    Currency:
-                </div>
-            </Grid>
-            <Grid
-                item
-                xs={9}
-            >
                 <Select
-                    label="Currency"
-                    name="currency"
                     style={{
-                        marginLeft: 15,
-                        width: '15%',
+                        width: '70%',
+                    }}
+                    label='Icon'
+                    name='icon'
+                    renderValue={(value) => {
+                        if (value === '') return <></>
+                        const icon = MUIIcons.filter(MUIIcon => MUIIcon.name === value)[0].icon
+                        return (
+                            createElement(icon, {key: icon}, null)
+                        )
                     }}
                     onChange={onInputChange}
-                    defaultValue={currencies[0]}
+                    defaultValue={editCategory ? getIconName(editCategory.icon) : ''}
                 >
-                    {currencies.map(currency => (
+                    {MUIIcons.map(MUIIcon => (
                         <MenuItem
-                            value={currency}
-                            key={currency}
+                            value={MUIIcon.name}
+                            key={MUIIcon.name}
                         >
-                            {currency}
+                            {createElement(MUIIcon.icon, {key: MUIIcon.icon}, null)}
                         </MenuItem>
                     ))}
                 </Select>
@@ -99,4 +89,4 @@ function AddBankAccountDialogContent({
     )
 }
 
-export default AddBankAccountDialogContent;
+export default EditCategoryDialogContent;
